+++ CHANGELOG.md
@@ -39,6 +39,8 @@ This release extends the planned support of the modules to OneView REST API vers
 - oneview_sas_logical_interconnect_group_facts
 - oneview_server_hardware
 - oneview_server_hardware_facts
+- oneview_server_hardware_type
+- oneview_server_hardware_type_facts
 - oneview_server_profile
 - oneview_server_profile_facts
 - oneview_server_profile_template